+++ kernel/string.c
@@ -234,7 +234,7 @@ void* memmove(void* dest_, const void* src_, size_t count) {
     if (dest > src && dest + count < src) {
         // overlap, src is lower.
         // move in reverse
-        for (size_t i=count-1; i>=0; i--) {
+        for (long i=count-1; i>=0; i--) {
             dest[i] = src[i];
         }
         return dest;
